# Fix to avoid adding too many nodes for unknown words